same changes as for 3.13.3

+++ src/rec/calcRecord.c
@@ -305,7 +305,7 @@ static void alarm(pcalc)
     struct calcRecord	*pcalc;
 {
 	double		val;
-	float		hyst, lalm, hihi, high, low, lolo;
+	double		hyst, lalm, hihi, high, low, lolo;
 	unsigned short	hhsv, llsv, hsv, lsv;
 
 	if(pcalc->udf == TRUE ){
